pyadcs_connector/requirements.txt
Django==5.2.4
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
decorator==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
gssapi==1.9.0
gunicorn==23.0.0
idna==3.10
krb5==0.7.1
packaging==25.0
pip==25.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pyspnego==0.11.2
pytz==2025.2
pywinrm==0.5.0
requests==2.32.4
requests-credssp==2.0.0
requests-kerberos==0.15.0
setuptools~=80.9.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
xmltodict==0.14.2